:bug: Fix Level Generator and sort everything
