+++ Assets/Scenes/LevelGeneratorTest.unity
@@ -108,106 +108,6 @@ NavMeshSettings:
     tileSize: 256
     accuratePlacement: 0
   m_NavMeshData: {fileID: 0}
---- !u!1001 &76619311
-Prefab:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0.7
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_RootOrder
-      value: 2
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 10
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalScale.y
-      value: 10
-      objectReference: {fileID: 0}
-    - target: {fileID: 400000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_LocalScale.z
-      value: 10
-      objectReference: {fileID: 0}
-    - target: {fileID: 100000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-      propertyPath: m_TagString
-      value: Player
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 100100000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-  m_IsPrefabParent: 0
---- !u!1001 &683651738
-Prefab:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalPosition.x
-      value: 3.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalPosition.y
-      value: 4.21
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalPosition.z
-      value: 11
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4989677723461180, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-      propertyPath: m_RootOrder
-      value: 4
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 100100000, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-  m_IsPrefabParent: 0
 --- !u!1 &720377549
 GameObject:
   m_ObjectHideFlags: 0
@@ -232,21 +132,27 @@ MonoBehaviour:
   m_GameObject: {fileID: 720377549}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 63c6b4ffb5c8345c9a4eaad00c0183e5, type: 3}
+  m_Script: {fileID: 11500000, guid: 55c8aaeb0963c4752903b2f3e954c556, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  map: {fileID: 2800000, guid: 3fc1ddc217f514140b47a29319950ef1, type: 3}
+  map: {fileID: 2800000, guid: 54a77c8377c4a4d7398b92f3f3cd8e4f, type: 3}
   tiles:
-  - color: {r: 0, g: 0, b: 0, a: 1}
-    prefab: {fileID: 1664977490947150, guid: cdee86336b89a44a1b3eeac3f150a653, type: 2}
-  - color: {r: 1, g: 0, b: 0, a: 1}
-    prefab: {fileID: 1635733682990848, guid: dcf05c9f9371c43d596aa984bdc4f704, type: 2}
-  - color: {r: 1, g: 1, b: 0, a: 1}
-    prefab: {fileID: 1606032963425016, guid: 3eb7ab82db468f44192450589439112c, type: 2}
-  - color: {r: 0, g: 1, b: 0, a: 1}
-    prefab: {fileID: 1464707701030772, guid: 4d2adf93fa37d7a48a02ad64183cafbd, type: 2}
-  - color: {r: 1, g: 0, b: 1, a: 1}
-    prefab: {fileID: 1709681267027154, guid: d3eaf64f610196f4cb48d2205fbc8800, type: 2}
+  - color:
+      serializedVersion: 2
+      rgba: 4278190080
+    prefab: {fileID: 1231044079733434, guid: 6a9d9bcadc5e6433e9023e0c40cc3c9b, type: 2}
+  - color:
+      serializedVersion: 2
+      rgba: 4278190335
+    prefab: {fileID: 1704960200295554, guid: 93e13c86aa44a46d0911beeb0e3f0c2d, type: 2}
+  - color:
+      serializedVersion: 2
+      rgba: 4278255615
+    prefab: {fileID: 1249022749346382, guid: 1377ac4f3784e4c15acc3dc7bb889cfa, type: 2}
+  - color:
+      serializedVersion: 2
+      rgba: 4278255360
+    prefab: {fileID: 1016175566669394, guid: c744530c4ca5a45e9b28ebf18eeb68a0, type: 2}
 --- !u!4 &720377551
 Transform:
   m_ObjectHideFlags: 0
@@ -337,7 +243,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 50, y: -30, z: 0}
 --- !u!1 &1739129800
 GameObject:
@@ -351,7 +257,6 @@ GameObject:
   - component: {fileID: 1739129803}
   - component: {fileID: 1739129802}
   - component: {fileID: 1739129801}
-  - component: {fileID: 1739129806}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -429,50 +334,3 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1739129806
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 1739129800}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6c946351b1354e34f92feff679656f60, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  myCamera: {fileID: 0}
-  duration: 0.3
-  magnitude: 0.25
---- !u!1 &2082384898 stripped
-GameObject:
-  m_PrefabParentObject: {fileID: 100000, guid: 3c1573b093600cb43b1cd498db9cefa5, type: 3}
-  m_PrefabInternal: {fileID: 76619311}
---- !u!70 &2082384901
-CapsuleCollider2D:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 2082384898}
-  m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0.09}
-  m_Size: {x: 0.08, y: 0.24}
-  m_Direction: 0
---- !u!114 &2082384902
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 2082384898}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 23829523e632a3b428f74654070f50bc, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  hasKey: 0
-  phone:
-    juice: 50
