--- Assets/Scripts/ColorToPrefab.cs
@@ -1,11 +1,0 @@
-﻿/*
-* Copyright (c) Mergers and Acupuncture
-*/
-using UnityEngine;
-
-[System.Serializable]
-public class ColorToPrefab
-{
-	public Color color;
-	public GameObject prefab;
-}
